src/test/javascript/spec/app/entities/veterinario/veterinario.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { VeterinarioService } from 'app/entities/veterinario/veterinario.service';
import { IVeterinario, Veterinario } from 'app/shared/model/veterinario.model';

describe('Service Tests', () => {
    describe('Veterinario Service', () => {
        let injector: TestBed;
        let service: VeterinarioService;
        let httpMock: HttpTestingController;
        let elemDefault: IVeterinario;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(VeterinarioService);
            httpMock = injector.get(HttpTestingController);

            elemDefault = new Veterinario(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 0, 0);
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign({}, elemDefault);
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a Veterinario', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .create(new Veterinario(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a Veterinario', async () => {
                const returnedFromService = Object.assign(
                    {
                        email: 'BBBBBB',
                        login: 'BBBBBB',
                        nome: 'BBBBBB',
                        senha: 'BBBBBB',
                        numeroRegistroCfmv: 1,
                        numeroRegistroCrmv: 1
                    },
                    elemDefault
                );

                const expected = Object.assign({}, returnedFromService);
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of Veterinario', async () => {
                const returnedFromService = Object.assign(
                    {
                        email: 'BBBBBB',
                        login: 'BBBBBB',
                        nome: 'BBBBBB',
                        senha: 'BBBBBB',
                        numeroRegistroCfmv: 1,
                        numeroRegistroCrmv: 1
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a Veterinario', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
